+++ include/dcc/rx/crtp_base.hpp
@@ -93,13 +93,14 @@ struct CrtpBase {
       impl().highCurrentBiDi(cv28 & ztl::make_mask(6u));
 
     // IDs
-    _did = {impl().readCv(250u - 1u),
-            impl().readCv(251u - 1u),
-            impl().readCv(252u - 1u),
-            impl().readCv(253u - 1u)};
+    _did = {impl().readCv(DCC_RX_LOGON_DID_CV_ADDRESS + 0u),
+            impl().readCv(DCC_RX_LOGON_DID_CV_ADDRESS + 1u),
+            impl().readCv(DCC_RX_LOGON_DID_CV_ADDRESS + 2u),
+            impl().readCv(DCC_RX_LOGON_DID_CV_ADDRESS + 3u)};
     _cids.front() = static_cast<decltype(_cids)::value_type>(
-      (impl().readCv(65297u - 1u) << 8u) | impl().readCv(65298u - 1u));
-    _session_ids.front() = impl().readCv(65299u - 1u);
+      (impl().readCv(DCC_RX_LOGON_CID_CV_ADDRESS + 0u) << 8u) |
+      impl().readCv(DCC_RX_LOGON_CID_CV_ADDRESS + 1u));
+    _sids.front() = impl().readCv(DCC_RX_LOGON_SID_CV_ADDRESS);
 
     // Initialization time point
     _tps.init = std::chrono::system_clock::now();
@@ -374,7 +375,7 @@ private:
   /// \retval true  Command accepted
   /// \retval false Command rejected
   bool executeAutomaticLogon(Address addr, std::span<uint8_t const> bytes) {
-    if (addr != 254u) return false;
+    if (!_logon_enabled || addr != 254u) return false;
 
     switch (bytes[0uz] & 0xF0u) {
       // SELECT
@@ -407,8 +408,8 @@ private:
       case 0b1111'0000u: {
         auto const gg{static_cast<AddressGroup>(bytes[0uz] & 0b11u)};
         auto const cid{data2uint16(&bytes[1uz])};
-        auto const session_id{bytes[3uz]};
-        logonEnable(gg, cid, session_id);
+        auto const sid{bytes[3uz]};
+        logonEnable(gg, cid, sid);
         break;
       }
     }
@@ -436,7 +437,8 @@ private:
     switch (bytes[0uz] & 0x0Fu) {
       case 0b0000'0010u: [[fallthrough]];
       case 0b0000'0011u:
-        if (!(bytes[1uz] & ztl::make_mask(7u)))
+        if (!(bytes[1uz] & ztl::make_mask(7u)) &&
+            _own_equal_packets_count == !DCC_STANDARD_COMPLIANCE + 1uz)
           cvWrite(19u - 1u,
                   static_cast<uint8_t>(bytes[0uz] << 7u | bytes[1uz]));
         break;
@@ -719,12 +721,10 @@ private:
       // Write byte
       case 0b11u:
         // Not enough packets
-        if (_own_equal_packets_count < DCC_RX_MIN_CV_WRITE_PACKETS)
+        if (_own_equal_packets_count < 2uz)
           ;
         // Write once
-        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS ||
-                 serviceMode())
-          cvWrite(cv_addr, bytes[2uz]);
+        else if (_own_equal_packets_count == 2uz) cvWrite(cv_addr, bytes[2uz]);
         // ...otherwise just verify
         else cvVerify(cv_addr, bytes[2uz]);
         break;
@@ -734,9 +734,7 @@ private:
         auto const pos{bytes[2uz] & 0b111u};
         auto const bit{static_cast<bool>(bytes[2uz] & ztl::make_mask(3u))};
         if (!(bytes[2uz] & ztl::make_mask(4u))) cvVerify(cv_addr, bit, pos);
-        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS ||
-                 serviceMode())
-          cvWrite(cv_addr, bit, pos);
+        else if (_own_equal_packets_count == 2uz) cvWrite(cv_addr, bit, pos);
         break;
       }
     }
@@ -761,19 +759,21 @@ private:
       // Acceleration adjustment (CV23)
       case 0b0010u:
         if (size(bytes) != 2uz + sizeof(_checksum)) return false;
-        else cvWrite(23u - 1u, bytes[1uz]);
+        else if (_own_equal_packets_count == !DCC_STANDARD_COMPLIANCE + 1uz)
+          cvWrite(23u - 1u, bytes[1uz]);
         break;
 
       // Deceleration adjustment (CV24)
       case 0b0011u:
         if (size(bytes) != 2uz + sizeof(_checksum)) return false;
-        else cvWrite(24u - 1u, bytes[1uz]);
+        else if (_own_equal_packets_count == !DCC_STANDARD_COMPLIANCE + 1uz)
+          cvWrite(24u - 1u, bytes[1uz]);
         break;
 
       // Extended address 0 and 1 (CV17 and CV18)
       case 0b0100u:
         if (size(bytes) != 3uz + sizeof(_checksum)) return false;
-        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS) {
+        else if (_own_equal_packets_count == 2uz) {
           cvWrite(17u - 1u, static_cast<uint8_t>(0b1100'0000u | bytes[1uz]));
           cvWrite(18u - 1u, bytes[2uz]);
           cvWrite(29u - 1u, true, 5u);
@@ -783,7 +783,7 @@ private:
       // Index high and index low (CV31 and CV32)
       case 0b0101u:
         if (size(bytes) != 3uz + sizeof(_checksum)) return false;
-        else if (_own_equal_packets_count == DCC_RX_MIN_CV_WRITE_PACKETS) {
+        else if (_own_equal_packets_count == 2uz) {
           cvWrite(31u - 1u, bytes[1uz]);
           cvWrite(32u - 1u, bytes[2uz]);
         }
@@ -1000,34 +1000,33 @@ private:
 
   /// Logon enable
   ///
-  /// \param  gg          Address group
-  /// \param  cid         Command station ID
-  /// \param  session_id  Session ID
-  void logonEnable(AddressGroup gg, uint16_t cid, uint8_t session_id) {
-    if (!_logon_enabled || _logon_selected) return;
+  /// \param  gg  Address group
+  /// \param  cid Command station ID
+  /// \param  sid Session ID
+  void logonEnable(AddressGroup gg, uint16_t cid, uint8_t sid) {
+    // Already got selected and CID/SID didn't change
+    if (_logon_selected && _cids.back() == cid && _sids.back() == sid) return;
+    // ...otherwise clear selected
+    else _logon_selected = false;
 
     // Store new CID and session ID
     _cids.back() = cid;
-    _session_ids.back() = session_id;
+    _sids.back() = sid;
 
-    // Exceptions
-    if (_cids.back() == _cids.front()) {
-      // Difference decides whether fast logon or skip
-      auto const session_id_diff{
-        static_cast<uint32_t>(_session_ids.back() - _session_ids.front())};
-      auto const skip{session_id_diff <= 4u};
-      _logon_selected = _logon_assigned = _logon_store = skip;
-
-      // Skip logon if diff between session IDs is <=4
-      if (skip) {
-        std::array const cv65300_65301{impl().readCv(65300u - 1u),
-                                       impl().readCv(65301u - 1u)};
-        _addrs.logon = decode_address(cv65300_65301);
-        return;
-      }
-      // Force new logon if diff is >4
-      else
-        _addrs.logon = 0u;
+    // Skip logon if CID equal and diff between session IDs <=1
+    if (auto const skip{_cids.back() == _cids.front() &&
+                        static_cast<uint8_t>(_sids.back() - _sids.front()) <=
+                          1u}) {
+      _logon_selected = _logon_assigned = _logon_store = true;
+      std::array const cv65300_65301{
+        impl().readCv(DCC_RX_LOGON_ADDRESS_CV_ADDRESS + 0u),
+        impl().readCv(DCC_RX_LOGON_ADDRESS_CV_ADDRESS + 1u)};
+      _addrs.logon = decode_address(cv65300_65301);
+      return;
+    }
+    // ...otherwise force new logon
+    else {
+      _addrs.logon = 0u;
     }
 
     switch (gg) {
@@ -1052,8 +1051,7 @@ private:
   ///
   /// \param  did Unique ID
   void logonSelect(std::span<uint8_t const, 4uz> did) {
-    if (!_logon_enabled || _logon_assigned || !std::ranges::equal(did, _did))
-      return;
+    if (_logon_assigned || !std::ranges::equal(did, _did)) return;
     _logon_selected = true;
     std::array<uint8_t, 5uz> data{
       static_cast<uint8_t>(ztl::make_mask(7u) | (_addrs.primary >> 8u)),
@@ -1079,7 +1077,7 @@ private:
   void logonAssign(std::span<uint8_t const, 4uz> did,
                    AddressAssign bb,
                    Address addr) {
-    if (!_logon_enabled || !std::ranges::equal(did, _did)) return;
+    if (!std::ranges::equal(did, _did)) return;
     _logon_assigned = _logon_store = true;
     _addrs.consist = 0u;
     _addrs.logon = addr;
@@ -1244,13 +1242,18 @@ private:
     impl().writeCv(20u - 1u, 0u);
 
     _cids.front() = _cids.back();
-    impl().writeCv(65297u - 1u, static_cast<uint8_t>(_cids.back() >> 8u));
-    impl().writeCv(65298u - 1u, static_cast<uint8_t>(_cids.back()));
+    impl().writeCv(DCC_RX_LOGON_CID_CV_ADDRESS + 0u,
+                   static_cast<uint8_t>(_cids.back() >> 8u));
+    impl().writeCv(DCC_RX_LOGON_CID_CV_ADDRESS + 1u,
+                   static_cast<uint8_t>(_cids.back()));
 
-    _session_ids.front() = _session_ids.back();
-    impl().writeCv(65299u - 1u, _session_ids.back());
-    impl().writeCv(65300u - 1u, static_cast<uint8_t>(_addrs.logon >> 8u));
-    impl().writeCv(65301u - 1u, static_cast<uint8_t>(_addrs.logon));
+    _sids.front() = _sids.back();
+    impl().writeCv(DCC_RX_LOGON_SID_CV_ADDRESS, _sids.back());
+
+    std::array<uint8_t, 2uz> cv65300_65301;
+    encode_address(_addrs.logon, begin(cv65300_65301));
+    impl().writeCv(DCC_RX_LOGON_ADDRESS_CV_ADDRESS + 0u, cv65300_65301[0uz]);
+    impl().writeCv(DCC_RX_LOGON_ADDRESS_CV_ADDRESS + 1u, cv65300_65301[1uz]);
   }
 
   /// Update quality of service (roughly every second)
@@ -1322,8 +1325,8 @@ private:
   Backoff _logon_backoff{};
   Backoff _tos_backoff{};
 
-  std::array<uint16_t, 2uz> _cids{};       ///< Central ID
-  std::array<uint8_t, 2uz> _session_ids{}; ///< Session ID
+  std::array<uint16_t, 2uz> _cids{}; ///< Central ID
+  std::array<uint8_t, 2uz> _sids{};  ///< Session ID
 
   uint8_t _qos{}; ///< Quality of service